refactor: extract application actions
